+++ components/o3-editorial-typography/stories/core/topic-tag.stories.tsx
@@ -3,6 +3,7 @@ import {TopicTag as TopicTagTsx} from '../../src/tsx/index';
 import * as StoryTemplates from '../story-templates';
 
 import '../../src/css/brands/core.css';
+import links from '@financial-times/o3-figma-sb-links';
 
 export default {
 	title: 'Core/o3-editorial-typography',
@@ -20,12 +21,10 @@ export default {
 	},
 } as Meta;
 
-const DesignParams = {
-	type: 'figma',
-	url: 'https://www.figma.com/file/5ATknbGociZMlnNXX4sy4f/Whitelabel---Design-System?type=design&node-id=4717-652&mode=design&t=Y50jCZbAtgxH2F3S-4',
-};
-
 export const TopicTag = StoryTemplates.TopicTag;
 TopicTag.parameters = {
-	design: DesignParams,
+	design: {
+		type: 'figma',
+		url: links['core-o3-editorial-typography--topic-tag'].figma
+	},
 };