+++ components/o3-editorial-typography/stories/whitelable/byline.stories.tsx
@@ -3,7 +3,7 @@ import {Byline as BylineTsx} from '../../src/tsx/index';
 import * as StoryTemplates from '../story-templates';
 
 import '../../src/css/brands/whitelabel.css';
-
+import links from '@financial-times/o3-figma-sb-links';
 export default {
 	title: 'Whitelabel/o3-editorial-typography',
 	component: BylineTsx,
@@ -17,19 +17,14 @@ export default {
 	parameters: {
 		backgrounds: {default: 'white'},
 		controls: {exclude: ['children', 'brand', 'theme']},
+		design: {
+			type: 'figma',
+			url: links['whitelabel-o3-editorial-typography--byline'].figma,
+		},
 	},
 } as Meta;
 
-const DesignParams = {
-	type: 'figma',
-	url: 'https://www.figma.com/file/5ATknbGociZMlnNXX4sy4f/Whitelabel---Design-System?type=design&node-id=4717-652&mode=design&t=Y50jCZbAtgxH2F3S-4',
-};
-
 export const Byline = StoryTemplates.Byline;
-Byline.parameters = {
-	design: DesignParams,
-};
-
 Byline.args = {
 	...StoryTemplates.Byline.args,
 	brand: 'whitelabel',