+++ components/o3-editorial-typography/stories/whitelable/caption.stories.tsx
@@ -3,7 +3,7 @@ import {Caption as CaptionTsx} from '../../src/tsx/index';
 import * as StoryTemplates from '../story-templates';
 
 import '../../src/css/brands/whitelabel.css';
-
+import links from '@financial-times/o3-figma-sb-links';
 export default {
 	title: 'Whitelabel/o3-editorial-typography',
 	component: CaptionTsx,
@@ -17,15 +17,13 @@ export default {
 	parameters: {
 		backgrounds: {default: 'white'},
 		controls: {exclude: ['children', 'theme']},
+		design: {
+			type: 'figma',
+			url: links['whitelabel-o3-editorial-typography--caption'].figma,
+		}
 	},
 } as Meta;
 
-const DesignParams = {
-	type: 'figma',
-	url: 'https://www.figma.com/file/5ATknbGociZMlnNXX4sy4f/Whitelabel---Design-System?type=design&node-id=4717-652&mode=design&t=Y50jCZbAtgxH2F3S-4',
-};
+
 
 export const Caption = StoryTemplates.Caption;
-Caption.parameters = {
-	design: DesignParams,
-};