+++ components/o3-tooltip/stories/sb-story-config.ts
@@ -5,7 +5,7 @@ export const TemplateSBConfig: Meta = {
 		html: {
 			root: '#component-wrapper',
 			transform: (code: string) => transformCode(code),
-		},
+		}
 	},
 };