+++ components/o3-typography/stories/core/professional/caption.stories.tsx
@@ -4,6 +4,7 @@ import * as TypographyStories from '../../story-templates';
 import '../../../main.css';
 
 import '../../../src/css/brands/professional.css';
+import links from '@financial-times/o3-figma-sb-links';
 
 export default {
 	title: 'Core/Professional/o3-typography/Caption',
@@ -17,6 +18,10 @@ export default {
 	],
 	parameters: {
 		backgrounds: {default: 'paper'},
+		design: {
+			type: 'figma',
+			url: links['core-professional-o3-typography-caption--caption-story'].figma,
+		}
 	},
 } as Meta;