+++ components/o3-typography/stories/core/professional/link.stories.tsx
@@ -2,6 +2,7 @@ import type {Meta} from '@storybook/react';
 
 import {Link, Body} from '../../../src/tsx';
 import '../../../src/css/brands/professional.css';
+import links from '@financial-times/o3-figma-sb-links';
 
 export default {
 	title: 'Core/Professional/o3-typography/Link',
@@ -15,6 +16,10 @@ export default {
 	],
 	parameters: {
 		backgrounds: {default: 'paper'},
+		design: {
+			type: 'figma',
+			url: links['core-professional-o3-typography-link--links'].figma,
+		},
 	},
 	argTypes: {
 		theme: {