+++ components/o3-typography/stories/core/professional/unorderedList.stories.tsx
@@ -2,6 +2,7 @@ import type {Meta} from '@storybook/react';
 import * as TypographyStories from '../../story-templates';
 import {UnorderedList} from '../../../src/tsx';
 import '../../../src/css/brands/professional.css';
+import links from '@financial-times/o3-figma-sb-links';
 
 export default {
 	title: 'Core/Professional/o3-typography/UnorderedList',
@@ -15,6 +16,12 @@ export default {
 	],
 	parameters: {
 		backgrounds: {default: 'paper'},
+		design: {
+			type: 'figma',
+			url: links[
+				'core-professional-o3-typography-unorderedlist--unordered-list-story'
+			].figma,
+		},
 	},
 } as Meta;