+++ components/o3-typography/stories/whitelabel/body.stories.tsx
@@ -4,6 +4,7 @@ import {Body as BodyTsx} from '../../src/tsx';
 import '../../src/css/brands/whitelabel.css';
 import {Meta} from '@storybook/react';
 
+import links from '@financial-times/o3-figma-sb-links';
 export default {
 	title: 'Whitelabel/o3-typography/Body',
 	component: BodyTsx,
@@ -16,6 +17,10 @@ export default {
 	],
 	parameters: {
 		backgrounds: {default: 'white'},
+		design: {
+			type: 'figma',
+			url: links['whitelabel-o3-typography-body--body'].figma,
+		}
 	},
 } as Meta;